LPD-71114 refactor: move all implementation from ai-hub-site-initializer module to ai-hub-impl
