+++ modules/dxp/apps/ai-hub/ai-hub-impl/bnd.bnd
@@ -1,0 +1,17 @@
+Bundle-Name: Liferay AI Hub Implementation
+Bundle-SymbolicName: com.liferay.ai.hub.impl
+Bundle-Version: 1.0.4
+Import-Package:\
+	!org.bouncycastle.jsse.*,\
+	\
+	!org.conscrypt.*,\
+	\
+	!org.openjsse.javax.net.ssl.*,\
+	!org.openjsse.net.ssl.*,\
+	\
+	android.*;resolution:=optional,\
+	\
+	*
+Web-ContextPath: /ai-hub-site-initializer
+-dsannotations-options: inherit
+-fixupmessages: Classes found in the wrong directory: ...;is:=ignore
