+++ modules/dxp/apps/ai-hub/ai-hub-impl/src/main/java/com/liferay/ai/hub/internal/assistant/handler/DefaultAssistantHandler.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.ai.hub.site.initializer.internal.assistant.handler;
+package com.liferay.ai.hub.internal.assistant.handler;
 
 import dev.langchain4j.invocation.InvocationParameters;
 import dev.langchain4j.service.AiServices;
